# swagger.yaml
...
openapi: 3.0.3
info:
  title: Clinic-AI API
  version: 1.0.0
  description: "AI-powered clinical intake system. This spec is being filled module-by-module."
servers:
  - url: http://localhost:8000
paths:
...
  /notes/transcribe:
    post:
      tags: [notes]
      summary: Transcribe audio
      description: Queue audio transcription and return immediately (202).
      operationId: transcribe_audio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [patient_id, visit_id, audio_file]
              properties:
                patient_id: {type: string}
                visit_id: {type: string}
                language: {type: string, default: en}
                audio_file: {type: string, format: binary}
      responses:
        '202': {description: Queued for transcription}
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Invalid audio file or visit status, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /notes/{patient_id}/visits/{visit_id}/transcript:
    get:
      tags: [notes]
      summary: Get transcript
      description: Get transcript for a visit.
      operationId: get_transcript
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Transcript returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TranscriptionSessionDTO'
        '404': {description: Transcript not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Invalid input, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /audio/:
    get:
      tags: [audio]
      summary: List audio files
      operationId: list_audio_files
      parameters:
        - name: patient_id
          in: query
          required: false
          schema: {type: string}
        - name: visit_id
          in: query
          required: false
          schema: {type: string}
        - name: audio_type
          in: query
          required: false
          schema: {type: string, description: 'adhoc | visit'}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          required: false
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: Audio files listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AudioListResponse'
        '500': {description: Failed to list audio files, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /transcription:
    post:
      tags: [transcription]
      summary: Adhoc transcription
      operationId: transcribe_audio_adhoc
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio_file]
              properties:
                audio_file: {type: string, format: binary}
                language: {type: string, default: en}
      responses:
        '200':
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AdhocTranscriptionResponse'
        '400': {description: No audio file or validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Invalid file type, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /transcription/structure:
    post:
      tags: [transcription]
      summary: Structure transcript text
      operationId: structure_transcript_text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StructureTextRequest'
      responses:
        '200': {description: Structured dialogue returned, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '400': {description: Empty transcript, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /transcription/adhoc/action-plan:
    post:
      tags: [transcription]
      summary: Queue action plan generation
      operationId: generate_action_plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionPlanRequest'
      responses:
        '202': {description: Action plan generation started, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_ActionPlanResponse'}}}}
        '404': {description: Adhoc not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: No transcript available, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /transcription/adhoc/{adhoc_id}/action-plan/status:
    get:
      tags: [transcription]
      summary: Get action plan status
      operationId: get_action_plan_status
      parameters:
        - name: adhoc_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Status returned, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /transcription/adhoc/{adhoc_id}/action-plan:
    get:
      tags: [transcription]
      summary: Get action plan
      operationId: get_action_plan
      parameters:
        - name: adhoc_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Action plan returned, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
components:
  schemas:
    ErrorResponse:
      type: object
      required: [success, error, message, timestamp, request_id]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: INTERNAL_SERVER_ERROR
        message:
          type: string
          example: An unexpected internal server error occurred.
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...
    ApiResponse_Dict:
      type: object
      required: [success, message, timestamp, request_id, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
        data:
          type: object
          additionalProperties: true
...
    TranscriptionSessionDTO:
      type: object
      required: [audio_file_path, transcript, transcription_status, started_at, completed_at, error_message, audio_duration_seconds, word_count]
      properties:
        audio_file_path: {type: string, nullable: true}
        transcript: {type: string, nullable: true}
        transcription_status: {type: string}
        started_at: {type: string, nullable: true}
        completed_at: {type: string, nullable: true}
        error_message: {type: string, nullable: true}
        audio_duration_seconds: {type: number, nullable: true}
        word_count: {type: integer, nullable: true}
        structured_dialogue:
          type: array
          items: {type: object, additionalProperties: true}
          nullable: true
    ApiResponse_TranscriptionSessionDTO:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: Success}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/TranscriptionSessionDTO'
...
    AudioFileResponse:
      type: object
      required: [audio_id, filename, content_type, file_size, audio_type, created_at, updated_at]
      properties:
        audio_id: {type: string}
        filename: {type: string}
        content_type: {type: string}
        file_size: {type: integer}
        duration_seconds: {type: number, nullable: true}
        patient_id: {type: string, nullable: true}
        visit_id: {type: string, nullable: true}
        adhoc_id: {type: string, nullable: true}
        audio_type: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
    AudioListResponse:
      type: object
      required: [files, total_count, limit, offset]
      properties:
        files:
          type: array
          items: {$ref: '#/components/schemas/AudioFileResponse'}
        total_count: {type: integer}
        limit: {type: integer}
        offset: {type: integer}
...
    ApiResponse_AudioListResponse:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: 'Audio files listed'}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/AudioListResponse'
...
    AdhocTranscriptionResponse:
      type: object
      required: [transcript]
      properties:
        transcript: {type: string}
        language: {type: string, nullable: true}
        confidence: {type: number, nullable: true}
        duration: {type: number, nullable: true}
        word_count: {type: integer, nullable: true}
        model: {type: string, nullable: true}
        filename: {type: string, nullable: true}
        adhoc_id: {type: string, nullable: true}
    StructureTextRequest:
      type: object
      required: [transcript]
      properties:
        transcript: {type: string}
        model: {type: string, nullable: true}
        adhoc_id: {type: string, nullable: true}
    ActionPlanRequest:
      type: object
      required: [adhoc_id]
      properties:
        adhoc_id: {type: string}
    ActionPlanResponse:
      type: object
      required: [adhoc_id, status, message]
      properties:
        adhoc_id: {type: string}
        status: {type: string}
        message: {type: string}
    ApiResponse_AdhocTranscriptionResponse:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: Success}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/AdhocTranscriptionResponse'
    ApiResponse_ActionPlanResponse:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: 'Action plan generation started'}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/ActionPlanResponse'
...